♻️Refactor(*): 프로젝트 구조 2차 변경
- 모든 폴더를 패키지로 인식시키기 위해서 __init__.py를 추가합니다.

[프로젝트 구조]
- domain: 도메인 엔티티, VO 정의. 상태와 행위를 함께 다루는 객체 중심
- repository: DB 쿼리 처리 계층. SQL 직접 사용 가능
- service: 유스케이스 로직. 여러 도메인/레포지토리를 조합하거나 트랜잭션 단위 구성
- router: FastAPI 라우팅. @router.get 등 API 외부 인터페이스 처리
- schema: 외부와의 입출력 데이터 정의. FastAPI Pydantic 모델들 (DTO)
